backend/requirements.txt
alembic==1.16.5
anyio==4.10.0
asyncpg==0.30.0
bcrypt==4.3.0
email-validator==2.3.0
fastapi==0.116.1
passlib==1.7.4
psycopg2-binary==2.9.10
pydantic==2.11.7
python-jose==3.5.0
requests==2.32.5
SQLAlchemy==2.0.43
starlette==0.47.3
typing_extensions==4.15.0
uvicorn==0.35.0
